+++ randomizer-core/src/test/java/de/alpharogroup/random/NullSecureRandomExtensionsTest.java
@@ -45,7 +45,6 @@ import java.util.Map;
 import java.util.Set;
 
 import org.apache.commons.lang3.ArrayUtils;
-import org.apache.log4j.Logger;
 import org.meanbean.test.BeanTestException;
 import org.meanbean.test.BeanTester;
 import org.testng.annotations.AfterMethod;
@@ -67,10 +66,6 @@ import de.alpharogroup.test.objects.enums.Gender;
  */
 public class NullSecureRandomExtensionsTest extends BaseTestCase
 {
-
-	/** The Constant logger. */
-	private static final Logger logger = Logger
-		.getLogger(NullSecureRandomExtensionsTest.class.getName());
 
 	boolean expected;
 
@@ -350,7 +345,7 @@ public class NullSecureRandomExtensionsTest extends BaseTestCase
 	public void testRandomByteArray()
 	{
 		final byte[] randomByteArray = RandomExtensions.randomByteArray(8);
-		logger.debug(new String(randomByteArray, Charset.forName("UTF-8")));
+		assertTrue(randomByteArray.length == 8);
 	}
 
 	/**
@@ -489,7 +484,6 @@ public class NullSecureRandomExtensionsTest extends BaseTestCase
 	@Test
 	public void testRandomIntInt()
 	{
-		logger.debug("Generate 100 secure random numbers:");
 		for (int i = 0; i < 100; i++)
 		{
 			final int randomInt = RandomExtensions.randomInt(5);
@@ -528,7 +522,6 @@ public class NullSecureRandomExtensionsTest extends BaseTestCase
 	@Test(enabled = true)
 	public void testRandomLongLong()
 	{
-		logger.debug("Generate 100 secure random numbers:");
 		for (int i = 0; i < 100; i++)
 		{
 			final long randomLong = RandomExtensions.randomLong(5l);